requirements.txt
discord
jishaku
python-dotenv